making hints on missing plugins #9
divide into two notifications, one for not installed plugin and another one for disabled (but installed) plugin.
does not yet search the marketplace for the first option
